backend/requirements.txt
# ════════════════════════════════════════════════════════════════════════════════
# REQUISITOS DE PYTHON - ANEXO DE DATOS BACKEND
# ════════════════════════════════════════════════════════════════════════════════
#
# INSTALACIÓN:
# pip install -r requirements.txt
#
# ════════════════════════════════════════════════════════════════════════════════

# Framework Web
Flask==2.3.3
Flask-CORS==4.0.0

# Base de Datos
mysql-connector-python==8.0.33

# Variables de Entorno
python-dotenv==1.0.0

# Utilidades
Werkzeug==2.3.7
Jinja2==3.1.2

# Seguridad (opcional pero recomendado)
cryptography==41.0.3
bcrypt==5.0.0

# Manejo de Excel y archivos
openpyxl==3.10.10
pandas==2.0.3
